luci-app-zerotier: fix generate nftables rules when there is more than one route
according to #504 , the .nft file will broke if the $ip_segment is a multi-line var. 

therefore, handle this scenario with `while ....read -r`.

I've tested in a running immortalwrt shell, that this can handle both a single-line and multi-line $ip_segement,  but feel free to comment/try.

P.S. it can also be done via `xargs -I {}` but I think the busybox in immortalwrt/openwrt does not have `xargs -I` enabled by default.